test(channel): split registry trust boundary regressions
